requirements.txt
curses-menu
npyscreen
logbook
python-openstacksdk
